Allow closing popups with right mouse button and "activate" key
instead of just Escape, which is a bit a hard to reach during
normal gameplay.
The activate key part of this change already existed but was commented out,
no idea why..
The right mouse button thing is now hardcoded - using it for activate
is default, but the player might reconfigure it, in which case it will
still close the window (and OTOH if they bind it to another mouse button,
that will not close the popup).. I found no good way to do this in a more
generic way, because each mouse button has its own callback (in contrast to
keyboard keys that have one generic function that's called with the key).
I guess it's good enough, certainly better than before ;)

fixes https://github.com/haekb/nolf1-modernizer/issues/31